Add Raspple II metadata for NOOBS
This is the metadata for NOOBS to present Raspple II as an installation
candidate.  You put these files in `os/Raspple_II/` in a NOOBS directory
and add boot.tar.xz and root.tar.xz among them.  You'll need to also
modify `partitions.json` to reference the uncompressed tarball and nominal
partition sizes.  There'll be an automated build process for this later,
but for now we'll be doing it by hand.

Note, `release_notes.txt` is presently blank.  We should change that.