scripts/*: Scripts for determining partition sizes
Right now we are hardcoding the values in partitions.json for the boot
and root tarballs, but that's not ideal.  In fact, it's not a good idea
because sooner or later someone's going to forget to update them and
it's going to cause problems.

I've written here a couple of small scripts--macros really--to determine
the values for partitions.json, both of the uncompressed .tar.xz files,
and calculating the nominal sizes for boot and root partitions either as
Raspbian seems to be doing it now for root, or as seems logical for
boot.

These scripts are not used yet and need some error checking.