gen_partitions_json.sh: Generates partitions.json
Rather than trying to sed this file from a template, it seemed better
just to write out the whole file, that way there's only one place that
the stuff affecting it needs to be edited.  JSON is not that complex,
especially in this case, so the loss of syntax highlighting is not a big
deal there.